increasing pression

+++ trajectory_planning/trajectory_planning.h
@@ -34,29 +34,29 @@ namespace kamaz {
 namespace hagen {
         class TrajectoryPlanning {
             public:
-                TrajectoryPlanning(float speed = 2);
+                TrajectoryPlanning(double speed = 2);
                 ~TrajectoryPlanning() = default;
 
-                bool generate_ts(std::vector<Eigen::VectorXf> path);
+                bool generate_ts(std::vector<Eigen::VectorXd> path);
                 void traj_opt7();
-                void save_status(std::vector<std::vector<Eigen::VectorXf>>
+                void save_status(std::vector<std::vector<Eigen::VectorXd>>
                         , std::string file_name);
 
-                void get_desired_state(float time
-                        , std::vector<Eigen::VectorXf>& states);
-                std::pair<float, int > closest(float value);
+                void get_desired_state(double time
+                        , std::vector<Eigen::VectorXd>& states);
+                std::pair<double, int > closest(double value);
 
-                void generate_target_trajectory(std::vector<Eigen::VectorXf>&  target_trajectory
-  , std::string trajectory_to_be_flown_file_name);
+                void generate_target_trajectory(std::vector<Eigen::VectorXd>&  target_trajectory
+                       , std::string trajectory_to_be_flown_file_name);
 
-                std::vector<float> time_segs;
+                std::vector<double> time_segs;
 
-                Eigen::MatrixXf way_points;
-                float total_time;
-                float max_speed;
-                Eigen::MatrixXf X;
-                Eigen::MatrixXf A;
-                Eigen::MatrixXf Y;
+                Eigen::MatrixXd way_points;
+                double total_time;
+                double max_speed;
+                Eigen::MatrixXd X;
+                Eigen::MatrixXd A;
+                Eigen::MatrixXd Y;
 
             private:
             };
